requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
executing==2.0.0
Flask==3.0.0
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.5.0
jupyter_core==5.4.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.8
openai==0.28.1
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.8
Werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
